WIP progress

+++ smart_contracts/contract/src/ext_ffi.rs
@@ -701,7 +701,10 @@ extern "C" {
         out_size: usize,
     ) -> i32;
     ///
-    pub fn casper_load_call_stack(call_stack_len_ptr: *mut u8, result_size_ptr: *mut u8) -> i32;
+    pub fn casper_load_call_stack(
+        call_stack_len_ptr: *mut usize,
+        result_size_ptr: *mut usize,
+    ) -> i32;
     /// Prints data directly to standard output on the host.
     ///
     /// # Arguments
